Implement semaphore
semaphore provides an atomic increment/decrement operation on a file,
with the return value being 0 if the value was incremented from 0 or
decremented to 0.

This provides a way to ensure that only a single instance of a daemon
is running at a given point of time, by starting the daemon if the
increment operation returns 0, and stopping the daemon if the
decrement operation returns 0. It does rely on all calls which increment
the count being followed by a corresponding call to decrement the
counter.