Add bk-escort (replaces pipekill)
bk-escort manages a single child process.
It provides a socket which can be used to terminate the child.
If a process connects to the socket, bk-escort will terminate the child
by sending a SIGTERM, and then sending a SIGKILL if the child does not
return within several seconds. Once the child has terminated, bk-escort
will write a single byte to the socket and exit. This allows a process
to connect to the socket, and then block until the child returns.

This implementation does not allow setting the timeout, rate-limiting
child restarts, detecting and handling permanent failures, or running
hooks on child start and stop.