Add a FIXME mentioning a possible bug in the locking code
state and semaphore both use a simple truncate and replace method for
writing to a file. However, this is problematic if we recieve a signal
during those operations or between them, as the result will be a broken
file. We can't easily change this due to the way we lock (the obvious
fix would be to create a new file and move it over the old one).
Blocking signals might work but the program still risks being
interrupted by a SIGKILL.