bh-escort: fix signal handling issues
bh-escort did not properly handle queued signals, since (on Linux at
least) only a single copy of a signal is ever queued. Instead of
assuming that we get all signals ever delivered, assume that there was
one, probably more.

For our interactive "kill twice to kill hard" behaviour this is OK,
since the user can just hit CTRL-C again, but could be surprising when
sending SIGKILL, so only deal with SIGINTs like that.