bh-*: support templated services
Templated services are specified using the service@target syntax,
where the service is the name of the service directory, and the
target is passed on the command line of the pre, post, and run
scripts.

This can be used to template services, eg for a getty service,
the service might accept the tty to start the getty on as the first
argument. To start the resulting service, a user would use the
`bh-start getty@/dev/tty0` or similar, which starts the getty
service on /dev/tty0.